igraph_rs

Type Alias igraph_attribute_type_t

pub type igraph_attribute_type_t = c_uint;

\typedef igraph_attribute_type_t \brief The possible types of the attributes.

Values of this enum are used by the attribute interface to communicate the type of an attribute to igraph’s C core. When igraph is integrated in a high-level language, the attribute type reported by the interface may not necessarily have to match the exact data type in the high-level language as long as the attribute interface can provide a conversion from the native high-level attribute value to one of the data types listed here. When the high-level data type is complex and has no suitable conversion to one of the atomic igraph attribute types (numeric, string or Boolean), the attribute interface should report the attribute as having an “object” type, which is ignored by the C core. See also \ref igraph_attribute_table_t.

\enumval IGRAPH_ATTRIBUTE_UNSPECIFIED Currently used internally as a “null value” or “placeholder value” in some algorithms. Attribute records with this type must not be passed to igraph functions. \enumval IGRAPH_ATTRIBUTE_NUMERIC Numeric attribute. \enumval IGRAPH_ATTRIBUTE_BOOLEAN Logical values, true or false. \enumval IGRAPH_ATTRIBUTE_STRING String attribute. \enumval IGRAPH_ATTRIBUTE_OBJECT Custom attribute type, to be used for special data types by client applications. The R and Python interfaces use this for attributes that hold R or Python objects. Usually ignored by igraph functions.