igraph_rs

Type Alias igraph_error_type_t

pub type igraph_error_type_t = c_uint;

\typedef igraph_error_type_t \brief Error code type. These are the possible values returned by \a igraph functions. Note that these are interesting only if you defined an error handler with \ref igraph_set_error_handler(). Otherwise the program is aborted and the function causing the error never returns.

\enumval IGRAPH_SUCCESS The function successfully completed its task. \enumval IGRAPH_FAILURE Something went wrong. You’ll almost never meet this error as normally more specific error codes are used. \enumval IGRAPH_ENOMEM There wasn’t enough memory to allocate on the heap. \enumval IGRAPH_PARSEERROR A parse error was found in a file. \enumval IGRAPH_EINVAL A parameter’s value is invalid. E.g. negative number was specified as the number of vertices. \enumval IGRAPH_EXISTS A graph/vertex/edge attribute is already installed with the given name. \enumval IGRAPH_EINVVID Invalid vertex ID, negative or too big. \enumval IGRAPH_EINVEID Invalid edge ID, negative or too big. \enumval IGRAPH_EINVMODE Invalid mode parameter. \enumval IGRAPH_EFILE A file operation failed. E.g. a file doesn’t exist, or the user has no rights to open it. \enumval IGRAPH_UNIMPLEMENTED Attempted to call an unimplemented or disabled (at compile-time) function. \enumval IGRAPH_DIVERGED A numeric algorithm failed to converge. \enumval IGRAPH_ARPACK An error happened inside a calculation implemented in ARPACK. The calculation involved is most likely an eigenvector-related calculation. \enumval IGRAPH_ENEGCYCLE Negative cycle detected while calculating shortest paths. \enumval IGRAPH_EINTERNAL Internal error, likely a bug in igraph. \enumval IGRAPH_EATTRCOMBINE Unimplemented attribute combination method for the given attribute type. \enumval IGRAPH_EOVERFLOW Integer or double overflow. \enumval IGRAPH_EUNDERFLOW Integer or double underflow. \enumval IGRAPH_ERWSTUCK Random walk got stuck. \enumval IGRAPH_ERANGE Maximum vertex or edge count exceeded. \enumval IGRAPH_ENOSOL Input problem has no solution.